+++ tests/unit/testsuite/Bolt/Boltpay/Helper/ApiTest.php
@@ -8,16 +8,6 @@ require_once('TestHelper.php');
 class Bolt_Boltpay_Helper_ApiTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * @var int|null
-     */
-    private static $productId = null;
-
-    /**
-     * @var Bolt_Boltpay_TestHelper|null
-     */
-    private $testHelper;
-
-    /**
      * @var Bolt_Boltpay_Helper_Api
      */
     private $currentMock;
@@ -33,10 +23,8 @@ class Bolt_Boltpay_Helper_ApiTest extends PHPUnit_Framework_TestCase
         $appStore = $this->app->getStore();
         $appStore->resetConfig();
 
-        $this->testHelper = null;
-
         $this->currentMock = $this->getMockBuilder('Bolt_Boltpay_Helper_Api')
-            ->setMethods(['verify_hook_secret', 'verify_hook_api', 'getBoltContextInfo', 'getItemImageUrl'])
+            ->setMethods(['verify_hook_secret', 'verify_hook_api', 'getBoltContextInfo'])
             ->enableOriginalConstructor()
             ->getMock();
 
@@ -55,82 +43,14 @@ class Bolt_Boltpay_Helper_ApiTest extends PHPUnit_Framework_TestCase
         $this->testBoltResponse = (object) $testBoltResponse;
     }
 
-    /**
-     * Generate dummy products for testing purposes
-     */
-    public static function setUpBeforeClass()
-    {
-        // Create some dummy product:
-        self::$productId = Bolt_Boltpay_ProductProvider::createDummyProduct('PHPUNIT_TEST_' . 1);
-    }
-
-    /**
-     * Delete dummy products after the test
-     */
-    public static function tearDownAfterClass()
-    {
-        Bolt_Boltpay_ProductProvider::deleteDummyProduct(self::$productId);
-    }
-
-    public function testBuildCart()
-    {
-        $this->testHelper = new Bolt_Boltpay_TestHelper();
-        $cart = $this->testHelper->addProduct(self::$productId, 2);
-
-        $_quote = $cart->getQuote();
-        $_quote->reserveOrderId();
-        $_items = $_quote->getAllItems();
-        $_multipage = true;
-        $item = $_items[0];
-        $product = $item->getProduct();
-        $imageUrl = 'no-image';
-
-        $this->currentMock->method('getItemImageUrl')
-            ->will($this->returnValue($imageUrl));
-
-        $expected = array (
-            'order_reference' => $_quote->getParentQuoteId(),
-            'display_id' => $_quote->getReservedOrderId()."|".$_quote->getId(),
-            'items' =>
-                array (
-                    0 =>
-                        array (
-                            'reference' => $_quote->getId(),
-                            'image_url' => $imageUrl,
-                            'name' => $item->getName(),
-                            'sku' => $item->getSku(),
-                            'description' => substr($product->getDescription(), 0, 8182) ?: '',
-                            'total_amount' => round($item->getCalculationPrice() * 100 * $item->getQty()),
-                            'unit_price' => round($item->getCalculationPrice() * 100),
-                            'quantity' => $item->getQty(),
-                            'type' => 'physical'
-                        ),
-                ),
-            'currency' => $_quote->getQuoteCurrencyCode(),
-            'discounts' => array (),
-            'total_amount' => round($_quote->getSubtotal() * 100),
-        );
-
-        $result = $this->currentMock->buildCart($_quote, $_items, $_multipage);
-
-        $this->assertEquals($expected, $result);
-    }
-
-    public function testStoreHasAllCartItems()
-    {
-        $this->testHelper = new Bolt_Boltpay_TestHelper();
-        $cart = $this->testHelper->addProduct(self::$productId, 2);
-
-        $result = $this->currentMock->storeHasAllCartItems($cart->getQuote());
-
-        $this->assertTrue($result);
-    }
-
     public function testIsResponseError()
     {
         $response = (object) $this->testBoltResponse;
 
-        $result = $this->currentMock->isResponseError($response);
+        $isResponseErrorMethod = new ReflectionMethod($this->currentMock, 'isResponseError');
+        $isResponseErrorMethod->setAccessible(true);
+
+        $result = $isResponseErrorMethod->invoke($this->currentMock, $response);
 
         $this->assertFalse($result);
     }
@@ -140,7 +60,10 @@ class Bolt_Boltpay_Helper_ApiTest extends PHPUnit_Framework_TestCase
         $this->testBoltResponse->errors = ['some_error_key' => 'some_error_message'];
         $response = (object) $this->testBoltResponse;
 
-        $result = $this->currentMock->isResponseError($response);
+        $isResponseErrorMethod = new ReflectionMethod($this->currentMock, 'isResponseError');
+        $isResponseErrorMethod->setAccessible(true);
+
+        $result = $isResponseErrorMethod->invoke($this->currentMock, $response);
 
         $this->assertTrue($result);
     }
@@ -150,7 +73,10 @@ class Bolt_Boltpay_Helper_ApiTest extends PHPUnit_Framework_TestCase
         $this->testBoltResponse->error_code = 10603;
         $response = (object) $this->testBoltResponse;
 
-        $result = $this->currentMock->isResponseError($response);
+        $isResponseErrorMethod = new ReflectionMethod($this->currentMock, 'isResponseError');
+        $isResponseErrorMethod->setAccessible(true);
+
+        $result = $isResponseErrorMethod->invoke($this->currentMock, $response);
 
         $this->assertTrue($result);
     }