+++ tests/unit/testsuite/Bolt/Boltpay/TestHelper.php
@@ -137,20 +137,25 @@ class Bolt_Boltpay_TestHelper
 
         return ("
             var json_cart = $jsonCart;
+            var json_hints = {};
             var quote_id = '{$immutableQuoteId}';
             var order_completed = false;
 
             BoltCheckout.configure(
                 json_cart,
-                $jsonHints,
+                json_hints,
                 {
                   check: function() {
-                    $checkCustom
-                    $onCheckCallbackAdmin
                     if (!json_cart.orderToken) {
-                        alert(json_cart.error);
+                        if (typeof BoltPopup !== \"undefined\") {
+                            BoltPopup.addMessage(json_cart.error).show();
+                        } else {
+                            alert(json_cart.error);
+                        }
                         return false;
                     }
+                    $checkCustom
+                    $onCheckCallbackAdmin
                     return true;
                   },