+++ CHANGELOG.md
@@ -1,4 +1,43 @@
 # Changelog
+## [v1.3.3](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.3.3.zip) 2018-10-02
+ - Fix for not including JavaScript when not needed
+ - Bugfix around shipping&tax
+
+## [v1.3.2](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.3.2.zip) 2018-09-29
+ - Refactoring
+ - Minor bugfix
+
+## [v1.3.1](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.3.1.zip) 2018-09-25
+ - Refactoring
+ - Fix issue with firecheckout
+ - Allow merchants to edit orders
+
+## [v1.3.0](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.3.0.zip) 2018-09-17
+ - Add cron to cleanup expired immutable quotes
+ - Remove JS alert and replaced with custom UI
+
+## [v1.2.3](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.2.3.zip) 2018-09-12
+ - Improved firecheckout loading performance
+ - Fix issue with webhook and user session
+ - Fix issue with missing region
+
+## [v1.2.2](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.2.2.zip) 2018-08-28
+ - Fixed off-by-one error
+ - Optimize loading
+
+## [v1.2.1](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.2.1.zip) 2018-08-17
+ - Introduced modman
+ - Tweaked explanation of publishable keys
+ - Bugfix for:
+  - shipping discount
+  - refund hook
+  - region_id
+
+## [v1.2.0](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.2.0.zip) 2018-07-29
+ - Bugfix around duplicate orders
+ - Admin order creation bugfix
+ - Better multi-store support
+
 ## [v1.1.7](https://s3-us-west-1.amazonaws.com/bolt-public/magento-integration-release/bolt-magento1_v1.1.7.zip) 2018-07-23
  - Add product type
  - Stability fix for quote lookup